auto-typer/requirements.txt
customtkinter>=5.2
pyautogui>=0.9.54
pyyaml>=6.0
httpx>=0.27
keyboard>=0.13
Pillow>=10.0
pyperclip>=1.8
